+++ server/src/controllers/schema.controller.js
@@ -5,7 +5,14 @@ var RestResult = require('../models/restResult.model');
 exports.hasSchema = function(req, res) {
     Schema.hasSchema()
       .then((result)=>{ res.json(new RestResult("success","schema and tables are ok",result.data?.success,result.data?.failed)) })
-      .catch((result)=>{ res.json(new RestResult("error","schema and tables are not ok",result.data?.success,result.data?.failed)) })
+      .catch((result)=>{ 
+        if(!result?.data){
+          res.json(new RestResult("error","FATAL ERROR",[],[result.message])) 
+        }else{
+          res.json(new RestResult("error","schema and tables are not ok",result.data?.success,result.data?.failed)) 
+        }
+        
+      })
 };
 exports.create = function(req, res) {
     Schema.create()